+++ blender_cloud/image_sharing.py
@@ -124,9 +124,8 @@ class PILLAR_OT_image_share(pillar.PillarOperatorMixin,
                 db_user = await self.check_credentials(context, REQUIRES_ROLES_FOR_IMAGE_SHARING)
                 self.user_id = db_user['_id']
                 self.log.debug('Found user ID: %s', self.user_id)
-            except pillar.NotSubscribedToCloudError:
-                self.log.exception('User not subscribed to cloud.')
-                self.report({'ERROR'}, 'Please subscribe to the Blender Cloud.')
+            except pillar.NotSubscribedToCloudError as ex:
+                self._log_subscription_needed(can_renew=ex.can_renew)
                 self._state = 'QUIT'
                 return
             except pillar.UserNotLoggedInError: